change uint64_t to int64_t

+++ cpp/src/distance/neighbors/specializations/ivfflat_extend_float_int64_t.cu
@@ -29,7 +29,8 @@ namespace raft::neighbors::ivf_flat {
                        raft::neighbors::ivf_flat::index<T, IdxT>* idx,                        \
                        raft::device_matrix_view<const T, IdxT, row_major> new_vectors,        \
                        std::optional<raft::device_vector_view<const IdxT, IdxT>> new_indices);
-RAFT_MAKE_INSTANCE(uint8_t, uint64_t);
+
+RAFT_MAKE_INSTANCE(float, int64_t);
 
 #undef RAFT_MAKE_INSTANCE
